// front-end/src/app/shared/models/Pacchetti.ts
export interface IPacchetti {
    id: number;
    nome: string;
    descrizione?: string;
    categoria: string; // PILATES, YOGA, MATWORK, etc.
    livello: string; // PRINCIPIANTE, INTERMEDIO, AVANZATO
    durataMinuti: number; // Durata in minuti
    maxPartecipanti: number;
    prezzo: number;
    attivo: boolean;
    dataCreazione: string;
    dataModifica?: string;
}